nba-matchup-predictor/nba-matchup-predictor/requirements.txt
apscheduler
apscheduler[sqlalchemy]
sqlalchemy
psycopg2-binary
pandas
numpy
scipy
scikit-learn
shap
joblib
filterpy
hmmlearn
statsmodels
requests
nba_api
python-dotenv
matplotlib
SQLAlchemy
APScheduler
